# Reduce legacy Windows driver ANSI escape sequence intermediate string allocations