## Fixes

- Improves #2725 for Windows Driver

## Proposed Changes

- [x] Add new methods to EscSeqUtils that append directly to StringBuilder.
- [x] Use the new methods in Windows Console and Driver.

## Pull Request checklist

- [ ] I've named my PR in the form of "Fixes #issue. Terse description."
- [x] My code follows the [style guidelines of Terminal.Gui](https://github.com/gui-cs/Terminal.Gui/blob/develop/.editorconfig) - if you use Visual Studio, hit `CTRL-K-D` to automatically reformat your files before committing.
- [x] My code follows the [Terminal.Gui library design guidelines](https://github.com/gui-cs/Terminal.Gui/blob/develop/CONTRIBUTING.md)
- [x] I ran `dotnet test` before commit
- [x] I have made corresponding changes to the API documentation (using `///` style comments)
- [x] My changes generate no new warnings
- [x] I have checked my code and corrected any poor grammar or misspellings
- [x] I conducted basic QA to assure all features are working

## Drivers

This should be applicable to the other drivers as well. I can of course make the changes if necessary but I have no WSL set up at the moment so I can only test the NetDriver.

## UICatalog Benchmark

Average time for me improved from ~22 sec to ~21 sec +/- 1 sec.

## Microbenchmarks

```

BenchmarkDotNet v0.14.0, Windows 10 (10.0.19045.5487/22H2/2022Update)
AMD Ryzen 7 5800X3D, 1 CPU, 16 logical and 8 physical cores
.NET SDK 8.0.406
  [Host]     : .NET 8.0.13 (8.0.1325.6609), X64 RyuJIT AVX2
  DefaultJob : .NET 8.0.13 (8.0.1325.6609), X64 RyuJIT AVX2

```
| Method | Mean      | Error    | StdDev   | Ratio | RatioSD | Gen0   | Allocated | Alloc Ratio |
|------- |----------:|---------:|---------:|------:|--------:|-------:|----------:|------------:|
| Set    | 162.82 ns | 2.867 ns | 2.394 ns |  1.00 |    0.02 | 0.0026 |     136 B |        1.00 |
| Append |  49.61 ns | 0.905 ns | 0.846 ns |  0.30 |    0.01 |      - |         - |        0.00 |

## Allocations

### Before

![1 before allocations](https://github.com/user-attachments/assets/2cd61ff8-dd32-470a-8dff-b44ff9d3b16a)

![1_1 escsequtils before allocations](https://github.com/user-attachments/assets/4db8558f-91b6-4ea9-bf79-71ee70d1fe4e)

### After

![2 after allocations](https://github.com/user-attachments/assets/165e9a03-1cb2-4972-a88d-036246b051f4)

![2_2 escsequtils after allocations](https://github.com/user-attachments/assets/1e58673d-689f-4252-a5dd-53620e2b2e2c)