@TheTonttu 2025-02-28
> The existing drivers should be replaced in due course by #3837 so may not be worth investing a lot of time optimising
> 
> But up to you. Changeover will only happen after period of both running side by side

Neat! I agree, seems a bit pointless to poke the current drivers any further then. Thankfully the new methods can also be utilized in the new drivers.